refactor(web): add useTickingClock hook + shared MM:SS formatter
Five panels (Lottery, CombatLog, Combat, Jukebox, MusicBox) each carried the
identical "const [now, setNow] = useState(() => Date.now())" plus a bare mount
interval to keep countdowns/relative timestamps fresh. Extract useTickingClock
and adopt it in exactly those unconditional sites — panels that gate the
interval on a condition (Quest/SkillBar/ActionBar cooldowns, Friends banners)
keep their own effect.

Also dedupes the byte-identical formatDuration(seconds)->M:SS helper that
JukeboxPanel and MusicBoxPanel both defined, into utils.formatMinutesSeconds.

Behaviour-preserving. tsc, eslint, and the web unit tests pass.
